Time: 6 ms (5.53%) | Memory: 27.4 MB (7.41%) - LeetSync

+++ LeetCode/Two_Pointers/42-trapping-rain-water/trapping-rain-water.cpp
@@ -8,7 +8,7 @@ public:
         partial_sum(suff_max.rbegin(), suff_max.rend(), suff_max.rbegin(), [](int a, int b) {
             return max(a, b);
         });
-        long long ans = 0;
+        int ans = 0;
         for(int i = 0; i < height.size(); i++) {
             ans += min(pref_max[i], suff_max[i]) - height[i];
         }
